Fix compilation
In https://cgit.FreeBSD.org/src/commit/?id=9ee759f3676f700f8224a95216f659f87f5d9ae9
the alignment of the structures used for the IPv4, IPv6, and
TCP header was reduced to 1. This causes the compilation to fail
due to -Wcast-align in combination with -Werror.
This could be worked around by using
CFLAGS += -Wno-error=cast-align
but I prefer to fix it.